fix name tag in board
