+++ src/app/utils/index.ts
@@ -1,6 +1,6 @@
 import { addQuestion } from "@services/client/question";
 import { useUserSession } from "@context/UserSessionContext";
-import { QuestionState } from "@interfaces/db";
+import { Question, QuestionState, TagOption } from "@interfaces/db";
 import { IdentifiableQuestion, IdentifiableQuestions } from "@interfaces/type";
 import { Timestamp } from "firebase/firestore";
 import { redirect } from "next/navigation";
@@ -69,40 +69,6 @@ export const formatTimeDifference = (
   return `${formattedHours}:${formattedMinutes} ${period}`;
 };
 
-export const defaultQuestion = () => {
-  return {
-    id: "",
-    title: "",
-    description: "",
-    questionPublic: false,
-    state: QuestionState.PENDING,
-    timestamp: Timestamp.now(),
-    group: [],
-    tags: [],
-  } as IdentifiableQuestion;
-};
-
-export const createQuestion = (
-  title: string,
-  description: string,
-  questionPublic: boolean,
-  timestamp: Timestamp,
-  group: string[],
-  tags: TagOption[],
-  courseId: string
-) => {
-  const question: Question = {
-    title: title,
-    description: description,
-    questionPublic: questionPublic,
-    state: QuestionState.PENDING,
-    timestamp: timestamp,
-    group: group,
-    tags: tags,
-  };
-
-  addQuestion(question, courseId);
-};
 /**
  * A question has "expired" if it's a different date or it's been resolved
  */
